Fix GLES3 and "present path fast" validation issue
Currently, if present path fast (aka RENDER-TO-BACKBUFFER) is turned on then it's not possible to create a GLES3 context. This is because present path fast makes every EGL config's 'CONFORMANT' flag = false, and there is some erroneous EGL validation logic in ANGLE that checks that configs are GLES3 conformant before allowing an app to create a GLES3 context.

The fix is to change the validation logic to check that the config supports *some* form of GLES3, rather than check if it's fully conformant.

Change-Id: I3acf9b74bf2eb78947c888c610659984293c2fd2